+++ packages/closed_household/lib/pages/closed_household_summary.dart
@@ -142,7 +142,7 @@ class ClosedHouseholdSummaryPageState
                                 i18.closeHousehold.closeHouseholdVillageName),
                             value: ClosedHouseholdSingleton()
                                 .boundary!
-                                .name
+                                .code
                                 .toString(),
                             labelFlex: 5,
                           ),