+++ packages/digit_data_model/lib/data/local_store/sql_store/sql_store.dart
@@ -178,6 +178,8 @@ class LocalSqlDataStore extends _$LocalSqlDataStore {
             try {
               await migrator.addColumn(
                   attendanceRegister, attendanceRegister.localityCode);
+              await migrator.addColumn(
+                  service, service.relatedClientReferenceId);
             } catch (e) {}
           }
         },