+++ packages/digit_data_model/lib/data/local_store/sql_store/sql_store.g.dart
@@ -21459,6 +21459,12 @@ class $ServiceTable extends Service with TableInfo<$ServiceTable, ServiceData> {
   late final GeneratedColumn<String> serviceDefId = GeneratedColumn<String>(
       'service_def_id', aliasedName, true,
       type: DriftSqlType.string, requiredDuringInsert: false);
+  static const VerificationMeta _relatedClientReferenceIdMeta =
+      const VerificationMeta('relatedClientReferenceId');
+  @override
+  late final GeneratedColumn<String> relatedClientReferenceId =
+      GeneratedColumn<String>('related_client_reference_id', aliasedName, true,
+          type: DriftSqlType.string, requiredDuringInsert: false);
   static const VerificationMeta _isActiveMeta =
       const VerificationMeta('isActive');
   @override
@@ -21577,6 +21583,7 @@ class $ServiceTable extends Service with TableInfo<$ServiceTable, ServiceData> {
         id,
         clientId,
         serviceDefId,
+        relatedClientReferenceId,
         isActive,
         accountId,
         additionalDetails,
@@ -21619,6 +21626,13 @@ class $ServiceTable extends Service with TableInfo<$ServiceTable, ServiceData> {
           _serviceDefIdMeta,
           serviceDefId.isAcceptableOrUnknown(
               data['service_def_id']!, _serviceDefIdMeta));
+    }
+    if (data.containsKey('related_client_reference_id')) {
+      context.handle(
+          _relatedClientReferenceIdMeta,
+          relatedClientReferenceId.isAcceptableOrUnknown(
+              data['related_client_reference_id']!,
+              _relatedClientReferenceIdMeta));
     }
     if (data.containsKey('is_active')) {
       context.handle(_isActiveMeta,
@@ -21727,6 +21741,9 @@ class $ServiceTable extends Service with TableInfo<$ServiceTable, ServiceData> {
           .read(DriftSqlType.string, data['${effectivePrefix}client_id'])!,
       serviceDefId: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}service_def_id']),
+      relatedClientReferenceId: attachedDatabase.typeMapping.read(
+          DriftSqlType.string,
+          data['${effectivePrefix}related_client_reference_id']),
       isActive: attachedDatabase.typeMapping
           .read(DriftSqlType.bool, data['${effectivePrefix}is_active']),
       accountId: attachedDatabase.typeMapping
@@ -21774,6 +21791,7 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
   final String? id;
   final String clientId;
   final String? serviceDefId;
+  final String? relatedClientReferenceId;
   final bool? isActive;
   final String? accountId;
   final String? additionalDetails;
@@ -21795,6 +21813,7 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
       {this.id,
       required this.clientId,
       this.serviceDefId,
+      this.relatedClientReferenceId,
       this.isActive,
       this.accountId,
       this.additionalDetails,
@@ -21821,6 +21840,10 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
     map['client_id'] = Variable<String>(clientId);
     if (!nullToAbsent || serviceDefId != null) {
       map['service_def_id'] = Variable<String>(serviceDefId);
+    }
+    if (!nullToAbsent || relatedClientReferenceId != null) {
+      map['related_client_reference_id'] =
+          Variable<String>(relatedClientReferenceId);
     }
     if (!nullToAbsent || isActive != null) {
       map['is_active'] = Variable<bool>(isActive);
@@ -21883,6 +21906,9 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
       serviceDefId: serviceDefId == null && nullToAbsent
           ? const Value.absent()
           : Value(serviceDefId),
+      relatedClientReferenceId: relatedClientReferenceId == null && nullToAbsent
+          ? const Value.absent()
+          : Value(relatedClientReferenceId),
       isActive: isActive == null && nullToAbsent
           ? const Value.absent()
           : Value(isActive),
@@ -21944,6 +21970,8 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
       id: serializer.fromJson<String?>(json['id']),
       clientId: serializer.fromJson<String>(json['clientId']),
       serviceDefId: serializer.fromJson<String?>(json['serviceDefId']),
+      relatedClientReferenceId:
+          serializer.fromJson<String?>(json['relatedClientReferenceId']),
       isActive: serializer.fromJson<bool?>(json['isActive']),
       accountId: serializer.fromJson<String?>(json['accountId']),
       additionalDetails:
@@ -21972,6 +22000,8 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
       'id': serializer.toJson<String?>(id),
       'clientId': serializer.toJson<String>(clientId),
       'serviceDefId': serializer.toJson<String?>(serviceDefId),
+      'relatedClientReferenceId':
+          serializer.toJson<String?>(relatedClientReferenceId),
       'isActive': serializer.toJson<bool?>(isActive),
       'accountId': serializer.toJson<String?>(accountId),
       'additionalDetails': serializer.toJson<String?>(additionalDetails),
@@ -21996,6 +22026,7 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
           {Value<String?> id = const Value.absent(),
           String? clientId,
           Value<String?> serviceDefId = const Value.absent(),
+          Value<String?> relatedClientReferenceId = const Value.absent(),
           Value<bool?> isActive = const Value.absent(),
           Value<String?> accountId = const Value.absent(),
           Value<String?> additionalDetails = const Value.absent(),
@@ -22018,6 +22049,9 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
         clientId: clientId ?? this.clientId,
         serviceDefId:
             serviceDefId.present ? serviceDefId.value : this.serviceDefId,
+        relatedClientReferenceId: relatedClientReferenceId.present
+            ? relatedClientReferenceId.value
+            : this.relatedClientReferenceId,
         isActive: isActive.present ? isActive.value : this.isActive,
         accountId: accountId.present ? accountId.value : this.accountId,
         additionalDetails: additionalDetails.present
@@ -22063,6 +22097,7 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
           ..write('id: $id, ')
           ..write('clientId: $clientId, ')
           ..write('serviceDefId: $serviceDefId, ')
+          ..write('relatedClientReferenceId: $relatedClientReferenceId, ')
           ..write('isActive: $isActive, ')
           ..write('accountId: $accountId, ')
           ..write('additionalDetails: $additionalDetails, ')
@@ -22085,27 +22120,29 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
   }
 
   @override
-  int get hashCode => Object.hash(
-      id,
-      clientId,
-      serviceDefId,
-      isActive,
-      accountId,
-      additionalDetails,
-      createdAt,
-      auditCreatedBy,
-      nonRecoverableError,
-      auditCreatedTime,
-      clientCreatedTime,
-      clientModifiedBy,
-      clientCreatedBy,
-      clientModifiedTime,
-      auditModifiedBy,
-      auditModifiedTime,
-      tenantId,
-      isDeleted,
-      rowVersion,
-      additionalFields);
+  int get hashCode => Object.hashAll([
+        id,
+        clientId,
+        serviceDefId,
+        relatedClientReferenceId,
+        isActive,
+        accountId,
+        additionalDetails,
+        createdAt,
+        auditCreatedBy,
+        nonRecoverableError,
+        auditCreatedTime,
+        clientCreatedTime,
+        clientModifiedBy,
+        clientCreatedBy,
+        clientModifiedTime,
+        auditModifiedBy,
+        auditModifiedTime,
+        tenantId,
+        isDeleted,
+        rowVersion,
+        additionalFields
+      ]);
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
@@ -22113,6 +22150,7 @@ class ServiceData extends DataClass implements Insertable<ServiceData> {
           other.id == this.id &&
           other.clientId == this.clientId &&
           other.serviceDefId == this.serviceDefId &&
+          other.relatedClientReferenceId == this.relatedClientReferenceId &&
           other.isActive == this.isActive &&
           other.accountId == this.accountId &&
           other.additionalDetails == this.additionalDetails &&
@@ -22136,6 +22174,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
   final Value<String?> id;
   final Value<String> clientId;
   final Value<String?> serviceDefId;
+  final Value<String?> relatedClientReferenceId;
   final Value<bool?> isActive;
   final Value<String?> accountId;
   final Value<String?> additionalDetails;
@@ -22158,6 +22197,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
     this.id = const Value.absent(),
     this.clientId = const Value.absent(),
     this.serviceDefId = const Value.absent(),
+    this.relatedClientReferenceId = const Value.absent(),
     this.isActive = const Value.absent(),
     this.accountId = const Value.absent(),
     this.additionalDetails = const Value.absent(),
@@ -22181,6 +22221,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
     this.id = const Value.absent(),
     required String clientId,
     this.serviceDefId = const Value.absent(),
+    this.relatedClientReferenceId = const Value.absent(),
     this.isActive = const Value.absent(),
     this.accountId = const Value.absent(),
     this.additionalDetails = const Value.absent(),
@@ -22204,6 +22245,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
     Expression<String>? id,
     Expression<String>? clientId,
     Expression<String>? serviceDefId,
+    Expression<String>? relatedClientReferenceId,
     Expression<bool>? isActive,
     Expression<String>? accountId,
     Expression<String>? additionalDetails,
@@ -22227,6 +22269,8 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
       if (id != null) 'id': id,
       if (clientId != null) 'client_id': clientId,
       if (serviceDefId != null) 'service_def_id': serviceDefId,
+      if (relatedClientReferenceId != null)
+        'related_client_reference_id': relatedClientReferenceId,
       if (isActive != null) 'is_active': isActive,
       if (accountId != null) 'account_id': accountId,
       if (additionalDetails != null) 'additional_details': additionalDetails,
@@ -22254,6 +22298,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
       {Value<String?>? id,
       Value<String>? clientId,
       Value<String?>? serviceDefId,
+      Value<String?>? relatedClientReferenceId,
       Value<bool?>? isActive,
       Value<String?>? accountId,
       Value<String?>? additionalDetails,
@@ -22276,6 +22321,8 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
       id: id ?? this.id,
       clientId: clientId ?? this.clientId,
       serviceDefId: serviceDefId ?? this.serviceDefId,
+      relatedClientReferenceId:
+          relatedClientReferenceId ?? this.relatedClientReferenceId,
       isActive: isActive ?? this.isActive,
       accountId: accountId ?? this.accountId,
       additionalDetails: additionalDetails ?? this.additionalDetails,
@@ -22308,6 +22355,10 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
     }
     if (serviceDefId.present) {
       map['service_def_id'] = Variable<String>(serviceDefId.value);
+    }
+    if (relatedClientReferenceId.present) {
+      map['related_client_reference_id'] =
+          Variable<String>(relatedClientReferenceId.value);
     }
     if (isActive.present) {
       map['is_active'] = Variable<bool>(isActive.value);
@@ -22372,6 +22423,7 @@ class ServiceCompanion extends UpdateCompanion<ServiceData> {
           ..write('id: $id, ')
           ..write('clientId: $clientId, ')
           ..write('serviceDefId: $serviceDefId, ')
+          ..write('relatedClientReferenceId: $relatedClientReferenceId, ')
           ..write('isActive: $isActive, ')
           ..write('accountId: $accountId, ')
           ..write('additionalDetails: $additionalDetails, ')
@@ -48019,6 +48071,7 @@ typedef $$ServiceTableInsertCompanionBuilder = ServiceCompanion Function({
   Value<String?> id,
   required String clientId,
   Value<String?> serviceDefId,
+  Value<String?> relatedClientReferenceId,
   Value<bool?> isActive,
   Value<String?> accountId,
   Value<String?> additionalDetails,
@@ -48042,6 +48095,7 @@ typedef $$ServiceTableUpdateCompanionBuilder = ServiceCompanion Function({
   Value<String?> id,
   Value<String> clientId,
   Value<String?> serviceDefId,
+  Value<String?> relatedClientReferenceId,
   Value<bool?> isActive,
   Value<String?> accountId,
   Value<String?> additionalDetails,
@@ -48084,6 +48138,7 @@ class $$ServiceTableTableManager extends RootTableManager<
             Value<String?> id = const Value.absent(),
             Value<String> clientId = const Value.absent(),
             Value<String?> serviceDefId = const Value.absent(),
+            Value<String?> relatedClientReferenceId = const Value.absent(),
             Value<bool?> isActive = const Value.absent(),
             Value<String?> accountId = const Value.absent(),
             Value<String?> additionalDetails = const Value.absent(),
@@ -48107,6 +48162,7 @@ class $$ServiceTableTableManager extends RootTableManager<
             id: id,
             clientId: clientId,
             serviceDefId: serviceDefId,
+            relatedClientReferenceId: relatedClientReferenceId,
             isActive: isActive,
             accountId: accountId,
             additionalDetails: additionalDetails,
@@ -48130,6 +48186,7 @@ class $$ServiceTableTableManager extends RootTableManager<
             Value<String?> id = const Value.absent(),
             required String clientId,
             Value<String?> serviceDefId = const Value.absent(),
+            Value<String?> relatedClientReferenceId = const Value.absent(),
             Value<bool?> isActive = const Value.absent(),
             Value<String?> accountId = const Value.absent(),
             Value<String?> additionalDetails = const Value.absent(),
@@ -48153,6 +48210,7 @@ class $$ServiceTableTableManager extends RootTableManager<
             id: id,
             clientId: clientId,
             serviceDefId: serviceDefId,
+            relatedClientReferenceId: relatedClientReferenceId,
             isActive: isActive,
             accountId: accountId,
             additionalDetails: additionalDetails,
@@ -48204,6 +48262,12 @@ class $$ServiceTableFilterComposer
       column: $state.table.serviceDefId,
       builder: (column, joinBuilders) =>
           ColumnFilters(column, joinBuilders: joinBuilders));
+
+  ColumnFilters<String> get relatedClientReferenceId =>
+      $state.composableBuilder(
+          column: $state.table.relatedClientReferenceId,
+          builder: (column, joinBuilders) =>
+              ColumnFilters(column, joinBuilders: joinBuilders));
 
   ColumnFilters<bool> get isActive => $state.composableBuilder(
       column: $state.table.isActive,
@@ -48308,6 +48372,12 @@ class $$ServiceTableOrderingComposer
       column: $state.table.serviceDefId,
       builder: (column, joinBuilders) =>
           ColumnOrderings(column, joinBuilders: joinBuilders));
+
+  ColumnOrderings<String> get relatedClientReferenceId =>
+      $state.composableBuilder(
+          column: $state.table.relatedClientReferenceId,
+          builder: (column, joinBuilders) =>
+              ColumnOrderings(column, joinBuilders: joinBuilders));
 
   ColumnOrderings<bool> get isActive => $state.composableBuilder(
       column: $state.table.isActive,