+++ packages/registration_delivery/lib/pages/beneficiary/record_past_delivery_details.dart
@@ -258,7 +258,8 @@ class RecordPastDeliveryDetailsPageState
                             localizations.translate(
                               i18.deliverIntervention.recordPastDeliveryDeatils,
                             ),
-                            style: DigitTheme.instance.mobileTheme.textTheme.displayMedium,
+                            style: DigitTheme
+                                .instance.mobileTheme.textTheme.displayMedium,
                           ),
                         ),
                         ...(futureTaskList?.asMap().entries.map((entry) {