+++ packages/registration_delivery/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.5+3
 
 * Added Eligibility checklist
+* Modified ServiceAttributes referenceId and ServiceModel clientId to IdGen and passing
+  relatedClientReference in additionalFields
 
 ## 1.0.5+2